UNet_train/requirements.txt
absl-py==1.4.0
addict==2.4.0
aenum==3.1.15
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
brevitas==0.9.1
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
coloredlogs==15.0.1
comm==0.1.4
cycler==0.11.0
Cython==3.0.3
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
dependencies==2.0.1
dill==0.3.6
distlib==0.3.6
entrypoints==0.4
exceptiongroup==1.1.2
fastjsonschema==2.18.1
filelock==3.12.2
flatbuffers==2.0.7
fonttools==4.38.0
fsspec==2023.1.0
fvcore==0.1.5.post20221221
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
huggingface-hub==0.15.1
humanfriendly==10.0
identify==2.5.24
idna==3.4
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmcv==2.0.1
mmdeploy==1.2.0
mmengine==0.8.2
mpmath==1.3.0
multiprocess==0.70.14
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.8
netron==6.8.8
networkx==2.6.3
ninja==1.11.1
nodeenv==1.8.0
nose==1.3.7
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.14.0
onnx2pytorch==0.4.1
onnxconverter-common==1.13.0
onnxruntime==1.14.1
onnxruntime-tools==1.7.0
opencv-python==4.7.0.68
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.1
pluggy==1.2.0
portalocker==2.7.0
pre-commit==2.21.0
prettytable==3.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.20.2
psutil==5.9.4
ptyprocess==0.7.0
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.4.0
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.0
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sympy==1.10.1
tabulate==0.9.0
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
tf2onnx==1.14.0
thop==0.0.31.post2001170342
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.1
torch-encoding==1.0.1
torchcam==0.3.2
torchgeometry==0.1.2
torchsummary==1.5.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.30.2
transformers-onnx==0.0.2
typing_extensions==4.7.1
urllib3==1.26.14
virtualenv==20.23.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.2
widgetsnbextension==4.0.9
wrapt==1.15.0
xmltodict==0.13.0
yacs==0.1.8
yapf==0.40.1
zipp==3.12.1
